Update dependencies and fix warnings (#132)
* QSortFilterProxyModel::invalidateFilter is scheduled for deprecation in Qt 6.13

* Fix comparison warning

* Update CI actions

* Simplify CMake

* HDPS -> ManiVault

* Remove some unused headers and usings

+++ src/Common.h
@@ -15,5 +15,6 @@ constexpr auto ult(E e) noexcept
 
 inline bool isColumnInModelIndexRange(const QModelIndex& topLeft, const QModelIndex& bottomRight, const std::uint32_t& column)
 {
-    return topLeft.column() <= column || bottomRight.column() >= column;
+    return static_cast<std::uint32_t>(topLeft.column()) <= column 
+    || static_cast<std::uint32_t>(bottomRight.column()) >= column;
 }
